Implemented abstract factories; using reflection to find the factories.
We're using abstract factories so we can just send the abstract factories
into the page creation method.

That means all we have to do is create a new factory and a new file template
class when we implement the filtered paginated template files feature.

We don't need to touch Main.java to add a new factory which returns FileTemplates, which you'll
then add to list of FileTemplates which you would then send to the page creation method.

Whether this was worth it, for having only three types of FileTemplate, is debatable!